requirements.txt
boto3
Flask
python-dotenv
opencv-python
numpy
gradio
google-generativeai
requests